+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@ The following items are new for Enterprise Chef 11.1.5 and/or are changes from p
 
 * [opscode-webui] Opscode WebUI has been removed in favor of the Manage Console add-on.
 * [private-chef-cookbooks] Introduce pluggable HA architecture as alternative to DRBD.
+* [private-chef-cookbooks] Add bifrost_sql_database uri to orgmapper.conf
 * [private-chef-ctl] Add a gather-logs command to create a tarball of
   important logs and system information.